refractored coordinate operations into own class --> TESTING WIP
