+++ common/__init__.py
@@ -1,4 +1,4 @@
-from ._compare_histograms import CompareHistorgrams
 from ._phenotype_framework import PhenotypeFramework
+from ._coordinate_ops import CoordinateOperations
 
-__all__ = ["CompareHistorgrams", "PhenotypeFramework"]
+__all__ = ["PhenotypeFramework", "CoordinateOperations"]
